Loops rendering more natural by removing excess \n
I've removed excess newlines that are caused inside of LoopRendering,
whereas a newline is found twice in the following template:

{%each%} -- first newline
  hello world
{%endeach%} -- second newline

Since there are technically newlines in the loop, they show up in the
output, which is frustrating because they are doubled by the fact that
the first and last newline layer in a loop.

This code added feels the least hacky in order to remove newlines only
when following or leading a loop expression.