Bump dependency versions and fix issue with code coverage
There was a huge issue with loading the karma-coverage plugin.  After
debugging, I noticed it was caused by calling `require` in the
karma/lib/plugin.js code.  The error is a weird, "Invalid date" exception.  The
fix was to simply pre-cache the module inside the Grunt config, so that it was
already loaded and simple lookup in the plugins code.

No idea why this happened.