Adds in getFilter/Partial methods
This fixes a previous regression where Combyne used to override child
partials with the parent's.  This wasn't a good idea, so I removed it in
0.7.0.  The regression appeared in NodeGit where we register all
partials at the top level and expect nested partials to find them.  I've
instead opt'd to store an internal reference to the parent in nested
partials and added getters for both partials and filters that can crawl
up the parent chain.