Ensure that conditional arguments are parsed valid
While working on a sample benchmarking app, I noticed that the following
was invalid:

```
{%each databases as database name%}{{database}}: {{name}}{%endeach%}
```

Using the the object:

```
{
  "database1": "online"
}
```

I was unable to get the template to render correctly.

After inspecting the `template.tree` object, I was able to discover that
the tokenizer had correctly found two cases of ASSIGNMENT the `as` in
database.

I updated the AST generation code to account for actual assignment and
properly adjust depending on what's passed.